+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased 0.3.0 - 2020-xx-yy
+- [CSE] Discovery supports "attributes + children" return content.
+
 ## 0.2.1 - 2020-03-06
 - [APPS] Fixed wrong originator handling for already registered AEs.
 - [APPS] Added persistent storage support for AEs.